Refactor BrowserPopup after recent changes
